cells/fulfillment-cell/emailservice/requirements.txt
grpcio==1.60.0
grpcio-health-checking==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.0
jinja2==3.1.2
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation-grpc==0.42b0